third_party/blink/web_tests/webaudio/AudioBufferSource/audiobuffersource-detune-modulation.html

<!DOCTYPE html>
<html>
  <head>
    <title>
      Test AudioBufferSourceNode's detune modulation
    </title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script src="../resources/audit-util.js"></script>
    <script src="../resources/audiobuffersource-testing.js"></script>
    <script src="../resources/buffer-loader.js"></script>
    <script src="../resources/audio-file-utils.js"></script>
  </head>
  <body>
    <script>
      const sampleRate = 44100;
      const duration = 0.25;

      const context =
          new OfflineAudioContext(1, sampleRate * duration, sampleRate);
      let referenceBuffer;

      promise_test(async () => {
        [referenceBuffer] = await loadBuffers(context, [
          'resources/audiobuffersource-detune-modulation-expected.wav'
        ]);
      }, 'load-reference');

      promise_test(async () => {
        // With this setting, the detune will be changing continuously and
        // repeatedly within the range of [-1200, 1200] around 440Hz, based on
        // the input from the oscillator.
        createSawtoothWithModulation(context, 'detune', 440, 1200);

        const renderedBuffer = await context.startRendering();
        const actual = renderedBuffer.getChannelData(0);
        const expected = referenceBuffer.getChannelData(0);

        // Compare two buffers with arbitrary (yet reasonable)
        // constraints. There parameters are determined by try bot
        // experiments.
        compareBuffersWithConstraints_W3CTH(actual, expected, {
          prefix: '',
          thresholdSNR: 93.302,
          thresholdDiffULP: 1.0391,
          thresholdDiffCount: 0,
          bitDepth: 16,
        });

        const filename =
            'resources/audiobuffersource-detune-modulation-actual.wav';
        if (downloadAudioBuffer(renderedBuffer, filename)) {
          assert_true(true, 'Saved reference file: ' + filename);
        }
      }, 'generate-verify');
    </script>
  </body>
</html>
